Change node exporter listen address and stop listening on tailnet
This change is due to the relization that it makes more sense to enable
prometheus in agent mode, implying that metrics ports don't need to be
exposed beyong localhost. This means that we use agent mode to push the
local metrics to the master prometheus and the master prometheus doesn't
need to scrape. Each node handles the metrics it needs to scrape